validation/requirements.txt
PyMuPDF==1.27.1
Pillow==10.4.0
openai>=1.0.0
sympy==1.13.0
tqdm==4.66.0
PyPDF2==3.0.1
